Refine structure
