+++ apps/monolith/test/integration/ContosoUniversity.Mvc.IntegrationTests/CoursesController/CreateEndpointsTests.cs
@@ -4,23 +4,17 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
 
-using Microsoft.AspNetCore.Antiforgery;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-
 using Xunit;
 
 public class CreateEndpointsTests :
     IClassFixture<TestsConfiguration>,
-    IClassFixture<NoAntiforgeryWebApplicationFactory>
+    IClassFixture<CustomWebApplicationFactory>
 {
     private readonly HttpClient _client;
 
     public CreateEndpointsTests(
         TestsConfiguration config,
-        NoAntiforgeryWebApplicationFactory factory)
+        CustomWebApplicationFactory factory)
     {
         factory.ClientOptions.BaseAddress = config.BaseAddressHttpsUrl;
         factory.ClientOptions.AllowAutoRedirect = true;
@@ -45,35 +39,4 @@ public class CreateEndpointsTests :
         // var content = await response.Content.ReadAsStringAsync();
         // content.Should().Contain("The field 'Credits' must be between 0 and 5.");
     }
-}
-
-public class NoAntiforgeryWebApplicationFactory : CustomWebApplicationFactory
-{
-    protected override void ConfigureWebHost(IWebHostBuilder builder)
-    {
-        base.ConfigureWebHost(builder);
-
-        builder.ConfigureServices(services =>
-        {
-            services.RemoveAll<IAntiforgery>();
-            services.AddTransient<IAntiforgery, NoOpAntiforgery>();
-        });
-    }
-}
-
-file class NoOpAntiforgery : IAntiforgery
-{
-    public AntiforgeryTokenSet GetAndStoreTokens(HttpContext httpContext) =>
-        new("test", "test", "test", "test");
-
-    public AntiforgeryTokenSet GetTokens(HttpContext httpContext) =>
-        new("test", "test", "test", "test");
-
-    public Task<bool> IsRequestValidAsync(HttpContext httpContext) =>
-        Task.FromResult(true);
-
-    public void SetCookieTokenAndHeader(HttpContext httpContext) { }
-
-    public Task ValidateRequestAsync(HttpContext httpContext) =>
-        Task.CompletedTask;
 }
